Split test_dwarf.c into a separate file. NFC
This allows the line and column numbers used in the test to
match that of an actual file under source control.
